# Improve propagate performance for big annotations