--- a/src/main/scala/firrtl/Compiler.scala
+++ b/src/main/scala/firrtl/Compiler.scala
@@ -7,7 +7,7 @@ import java.io.Writer
 
 import firrtl.RenameMap.{CircularRenameException, IllegalRenameException}
 
-import scala.collection.mutable
+import scala.collection.{mutable, immutable}
 import firrtl.annotations._
 import firrtl.ir.{Circuit, Expression}
 import firrtl.Utils.{error, throwInternalError}
@@ -225,36 +225,37 @@ abstract class Transform extends LazyLogging {
       resAnno: AnnotationSeq,
       renameOpt: Option[RenameMap]): AnnotationSeq = {
     val newAnnotations = {
-      val inSet = mutable.LinkedHashSet() ++ inAnno
-      val resSet = mutable.LinkedHashSet() ++ resAnno
-      val deleted = (inSet -- resSet).map {
+      val inSet = immutable.ListSet() ++ inAnno
+      val deleted = (inSet -- resAnno).map {
         case DeletedAnnotation(xFormName, delAnno) => DeletedAnnotation(s"$xFormName+$name", delAnno)
         case anno => DeletedAnnotation(name, anno)
-      }
-      val created = resSet -- inSet
-      val unchanged = resSet & inSet
-      (deleted ++ created ++ unchanged)
+      }.toSeq
+      (deleted ++ resAnno)
+    }
+
+    val renames = renameOpt.getOrElse(RenameMap())
+
+    logger.debug {
+      val remapped2original = mutable.LinkedHashMap[Annotation, mutable.LinkedHashSet[Annotation]]()
+      val keysOfNote = mutable.LinkedHashSet[Annotation]()
+      newAnnotations.flatMap { anno =>
+        val remappedAnnos = anno.update(renames)
+        remappedAnnos.foreach { remapped =>
+          val set = remapped2original.getOrElseUpdate(remapped, mutable.LinkedHashSet.empty[Annotation])
+          set += anno
+          if(set.size > 1) keysOfNote += remapped
+        }
+        remappedAnnos
+      }.toSeq
+      keysOfNote.map( key =>
+        s"""|The following original annotations are renamed to the same new annotation.
+            |Original Annotations:\n  ${remapped2original(key).mkString("\n  ")}
+            |New Annotation:\n  $key""".stripMargin
+      ).mkString
     }
 
     // For each annotation, rename all annotations.
-    val renames = renameOpt.getOrElse(RenameMap())
-    val remapped2original = mutable.LinkedHashMap[Annotation, mutable.LinkedHashSet[Annotation]]()
-    val keysOfNote = mutable.LinkedHashSet[Annotation]()
-    val finalAnnotations = newAnnotations.flatMap { anno =>
-      val remappedAnnos = anno.update(renames)
-      remappedAnnos.foreach { remapped =>
-        val set = remapped2original.getOrElseUpdate(remapped, mutable.LinkedHashSet.empty[Annotation])
-        set += anno
-        if(set.size > 1) keysOfNote += remapped
-      }
-      remappedAnnos
-    }.toSeq
-    keysOfNote.foreach { key =>
-      logger.debug(s"""The following original annotations are renamed to the same new annotation.""")
-      logger.debug(s"""Original Annotations:\n  ${remapped2original(key).mkString("\n  ")}""")
-      logger.debug(s"""New Annotation:\n  $key""")
-    }
-    finalAnnotations
+    newAnnotations.flatMap(_.update(renames))
   }
 }