When passing around big annotations (e.g., the `FirrtlCircuitAnnotation` that #920 adds), the overhead of copying annotations in `Compiler.propagateAnnotations` explodes. This does three things to improve the performance if you have big annotations:

1. The `mutable.LinkedHashSet` introduced by #855 is changed to an `immutable.ListSet`. I'm not sure exactly what's going on here, but construction of `LinkedHashSet` seems to be approximately 6x slower than `ListSet`. :woman_shrugging: No mutable operations are used here, so this is a straightforward change. The ordered requirement of #855 is still preserved.

2. Any construction of `ListSet[Annotation]` is limited as much as possible. This removes the unnecessary construction of `created` and `unchangedSet`.

3. Some `LinkedHashSet` that were only used for debugging (i.e., used inside of `logger.debug`) are moved to live inside of a `logger.debug` scope.

#920, without this PR, results in a dramatic runtime increase for large designs. E.g., Rocket Chip `DefaultConfig` takes 16s to get through `ChirrtlToHighFirrtl`:

```
Time: 15187.3 ms
Form: HighForm
======== Finished Transform ChirrtlToHighFirrtl ========
```

#920, with this PR, gets this down to 2s:

```
Time: 1840.0 ms
Form: HighForm
======== Finished Transform ChirrtlToHighFirrtl ========
```

### Alternatives 

There are two main alternatives:

1. A `FirrtlCircuitAnnotation` (and other big annotations) shouldn't be passed around during lowering. This is a valid point and I plan to update #920 in this way. However, the problem outlined here will come up for any massive annotation that people want to move around.

2. The debug-only code could be removed outright. I leave that decision up to @azidar. If left in, then debug will be substantially slower than non-debug. This seems fine generally, but this is really crawling... There may be a further refactor here that would remove the use of `LinkedHashSet`, use something faster, or rely on (1) to get rid of the big offender here.