@jackkoenig 2018-11-27
I don't think it's the `copying` of annotations like `FirrtlCircuitAnnotation`, I think it's the hashing process. Thus every time you have to rehash it's going to be pretty expensive.

I suspect that the performance gains come entirely from never having to hash the `FirrtlCircuitAnnotation` in a `ListSet` (since insertion and lookup is a linear search, no hashing. But for larger numbers of annotations I think we'll run into some pretty big problems using `ListSet`. In particular, I think the `(inSet -- resAnno)` is be `O(n^2)` for `ListSet` while merely `O(n)` for `LinkedHashMap`.

All of this being said, benchmarking is king. Have you benchmarked just using the default stuff in rocket-chip or have you run it with a larger number of annotations? I can run experiments with the large number of annotations we use internally tomorrow. We should probably create a more formal benchmarking suite including lots of annotations ([here's the script I'm currently using, but it doesn't do any annotations at the moment](https://github.com/jackkoenig/firrtlbench/blob/master/benchmark.py)).