@seldridge 2018-11-27
Great insight, @jackkoenig. That makes much more sense. It should be the repeated hashing of a ~10MB circuit that's doing it. And yes `ListSet` won't work for sets with large numbers of elements... 

I've only benchmarked with `DefaultConfig` rocket (~300 annotations), so this like will not scale to large numbers of annotations. 

What would you consider a "large" number of annotations?

This obviously has to be a set backed by a hash to support large numbers of annotations. Reverting #855 to move away from the linked nature shouldn't change anything either. 

I can dig into this more... It would seem like this must use a set backed by a hash. However, that hash should be a (`lazy`?) `val` for annotations storing large things and should be copied if no input parameters change.