@seldridge 2018-11-27
Not a bug and not a complicated solution either. Just `lazy val` caching the `hashCode` in big annotations is sufficient. :man_facepalming: